Summary: Build an application that can allow users to build a roster. One the user is done it can print out the players names and their numbers starting from 1.
TODO:

Create an empty list to maintain the player names

Ask the user if they'd like to add player to the list.

If the user answer "Yes", let them type in a name and add it to the list.
If the user answer "No", print out the team 'roster'
print the number of players on the team

Print the player number and the player name *The player number should start at the number one

Select a goalkeeper from the above roster

Print the goalkeeper's name' *Remember that lists use a zero based index